my fixes

+++ daos/interfaces/userDao.go
@@ -4,6 +4,6 @@ import "OnlineTestGo/models"
 
 //UserDao interface to be implemented by userdaoimpl
 type UserDao interface {
-	SaveNewUser(u models.User) (int, error)
-        CheckUser(u models.User)(int,error)
+	SaveNewUser(u models.User) (int64, error)
+	CheckUser(u models.User) (int64, error)
 }
